Split latest_datamodel in an internal and external version.
