+++ components/server/src/external/database/datamodels.py
@@ -2,9 +2,9 @@
 
 from typing import Any
 
+import pymongo
 from pymongo.database import Database
 
-from database.datamodels import latest_datamodel
 from server_utilities.functions import iso_timestamp
 
 
@@ -53,3 +53,11 @@ def insert_new_datamodel(database: Database, data_model):
         del data_model["_id"]
     data_model["timestamp"] = iso_timestamp()
     return database.datamodels.insert_one(data_model)
+
+
+def latest_datamodel(database: Database, max_iso_timestamp: str = ""):
+    """Return the latest data model."""
+    timestamp_filter = dict(timestamp={"$lte": max_iso_timestamp}) if max_iso_timestamp else None
+    if data_model := database.datamodels.find_one(timestamp_filter, sort=[("timestamp", pymongo.DESCENDING)]):
+        data_model["_id"] = str(data_model["_id"])
+    return data_model or {}
