+++ components/server/src/external/routes/datamodel.py
@@ -3,8 +3,7 @@
 import bottle
 from pymongo.database import Database
 
-from database.datamodels import latest_datamodel
-
+from ..database.datamodels import latest_datamodel
 from ..utils.functions import md5_hash, report_date_time
