+++ components/server/src/external/routes/measurement.py
@@ -8,7 +8,6 @@ from typing import cast
 import bottle
 from pymongo.database import Database
 
-from database.datamodels import latest_datamodel
 from database.measurements import insert_new_measurement, latest_measurement
 from database.reports import latest_reports
 from server_utilities.type import MetricId, SourceId
@@ -16,6 +15,7 @@ from shared.model.measurement import Measurement
 from shared.model.metric import Metric
 
 from ..database import sessions
+from ..database.datamodels import latest_datamodel
 from ..database.measurements import count_measurements, measurements_by_metric
 from ..model.data import SourceData
 from ..utils.functions import report_date_time
