+++ components/server/src/internal/database/reports.py
@@ -4,8 +4,9 @@ from pymongo.database import Database
 
 from server_utilities.type import MetricId
 from shared.model.metric import Metric
-from database.datamodels import latest_datamodel
 from database.reports import latest_reports
+
+from .datamodels import latest_datamodel
 
 
 def latest_metric(database: Database, metric_uuid: MetricId) -> Metric | None:
